move tabs to the middle of the page
