+++ app/assets/javascripts/discourse/tests/unit/controllers/admin-customize-themes-show-index-test.js
@@ -7,52 +7,6 @@ module(
   function (hooks) {
     setupTest(hooks);
 
-    test("can display source url for remote themes", function (assert) {
-      const repoUrl = "https://github.com/discourse/discourse-brand-header.git";
-      const remoteTheme = Theme.create({
-        id: 2,
-        default: true,
-        name: "default",
-        remote_theme: {
-          remote_url: repoUrl,
-        },
-      });
-
-      const controller = this.owner.lookup(
-        "controller:admin-customize-themes-show-index"
-      );
-      controller.setProperties({ model: remoteTheme });
-
-      assert.deepEqual(
-        controller.remoteThemeLink,
-        repoUrl,
-        "returns theme's repo URL"
-      );
-    });
-
-    test("can display source url for remote theme branches", function (assert) {
-      const remoteTheme = Theme.create({
-        id: 2,
-        default: true,
-        name: "default",
-        remote_theme: {
-          remote_url: "https://github.com/discourse/discourse-brand-header.git",
-          branch: "beta",
-        },
-      });
-
-      const controller = this.owner.lookup(
-        "controller:admin-customize-themes-show-index"
-      );
-      controller.setProperties({ model: remoteTheme });
-
-      assert.deepEqual(
-        controller.remoteThemeLink,
-        "https://github.com/discourse/discourse-brand-header/tree/beta",
-        "returns theme's repo URL to branch"
-      );
-    });
-
     test("displays settings editor button with settings", function (assert) {
       const theme = Theme.create({
         id: 2,
